requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
altair==4.2.0
async-timeout==4.0.2
attrs==22.2.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.2
backoff==1.11.1
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.4
decorator==5.1.1
dill==0.3.6
docker==6.0.1
docopt==0.6.2
elasticsearch==7.17.8
entrypoints==0.4
farm-haystack==1.12.2
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.1
gunicorn==20.1.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.2.0
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langdetect==1.0.9
llvmlite==0.39.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
mlflow==2.1.1
mmh3==3.0.0
monotonic==1.6
more-itertools==9.0.0
msrest==0.7.1
multidict==6.0.4
networkx==3.0
nltk==3.8.1
num2words==0.5.12
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openai==0.26.1
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
posthog==2.2.0
protobuf==3.20.3
pyarrow==10.0.1
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
quantulum3==0.8.1
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.1.0
scikit-learn==1.2.0
scipy==1.10.0
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
streamlit==1.12.0
tabulate==0.9.0
threadpoolctl==3.1.0
tika==2.6.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
websocket-client==1.4.2
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
